frontend/requirements.txt
gradio>=4.0.0
requests>=2.28.0
python-docx>=1.1.0
Pillow>=10.0.0